put yaml config in separate file
