+++ traj_opt/examples/example.h
@@ -1,0 +1,7 @@
+#pragma once
+
+/**
+ * Abstract base class for trajectory optimization examples.
+ */
+class TrajOptExample {
+}
