+++ traj_opt/examples/yaml_config.h
@@ -1,0 +1,70 @@
+#pragma once
+
+#include <string>
+#include <vector>
+
+#include "drake/common/yaml/yaml_io.h"
+
+namespace drake {
+namespace traj_opt {
+namespace examples {
+
+/**
+ * A simple object which stores parameters that define an optimization problem
+ * and various options, and can be loaded from a YAML file.
+ *
+ * See, e.g., spinner.yaml for an explanation of each field, and
+ * https://drake.mit.edu/doxygen_cxx/group__yaml__serialization.html for details
+ * on loading options from YAML.
+ */
+struct TrajOptExampleParams {
+  template <typename Archive>
+  void Serialize(Archive* a) {
+    a->Visit(DRAKE_NVP(q_init));
+    a->Visit(DRAKE_NVP(v_init));
+    a->Visit(DRAKE_NVP(q_nom));
+    a->Visit(DRAKE_NVP(v_nom));
+    a->Visit(DRAKE_NVP(q_guess));
+    a->Visit(DRAKE_NVP(Qq));
+    a->Visit(DRAKE_NVP(Qv));
+    a->Visit(DRAKE_NVP(R));
+    a->Visit(DRAKE_NVP(Qfq));
+    a->Visit(DRAKE_NVP(Qfv));
+    a->Visit(DRAKE_NVP(time_step));
+    a->Visit(DRAKE_NVP(num_steps));
+    a->Visit(DRAKE_NVP(max_iters));
+    a->Visit(DRAKE_NVP(linesearch));
+    a->Visit(DRAKE_NVP(play_optimal_trajectory));
+    a->Visit(DRAKE_NVP(play_initial_guess));
+    a->Visit(DRAKE_NVP(linesearch_plot_every_iteration));
+    a->Visit(DRAKE_NVP(print_debug_data));
+    a->Visit(DRAKE_NVP(F));
+    a->Visit(DRAKE_NVP(delta));
+    a->Visit(DRAKE_NVP(n));
+  }
+  std::vector<double> q_init;
+  std::vector<double> v_init;
+  std::vector<double> q_nom;
+  std::vector<double> v_nom;
+  std::vector<double> q_guess;
+  std::vector<double> Qq;
+  std::vector<double> Qv;
+  std::vector<double> R;
+  std::vector<double> Qfq;
+  std::vector<double> Qfv;
+  double time_step;
+  int num_steps;
+  int max_iters;
+  std::string linesearch;
+  bool play_optimal_trajectory;
+  bool play_initial_guess;
+  bool linesearch_plot_every_iteration;
+  bool print_debug_data;
+  double F;
+  double delta;
+  double n;
+};
+
+}  // namespace examples
+}  // namespace traj_opt
+}  // namespace drake
